+++ kmmbridge/src/main/kotlin/co/touchlab/kmmbridge/artifactmanager/GithubReleaseArtifactManager.kt
@@ -1,9 +1,9 @@
-package co.touchlab.faktory.artifactmanager
+package co.touchlab.kmmbridge.artifactmanager
 
-import co.touchlab.faktory.internal.GithubCalls
-import co.touchlab.faktory.internal.githubPublishToken
-import co.touchlab.faktory.internal.githubRepo
-import co.touchlab.faktory.kmmBridgeExtension
+import co.touchlab.kmmbridge.internal.GithubCalls
+import co.touchlab.kmmbridge.internal.githubPublishToken
+import co.touchlab.kmmbridge.internal.githubRepo
+import co.touchlab.kmmbridge.kmmBridgeExtension
 import org.gradle.api.GradleException
 import org.gradle.api.Project
 import org.gradle.api.Task
