+++ kmmbridge/src/main/kotlin/co/touchlab/kmmbridge/artifactmanager/MavenPublishArtifactManager.kt
@@ -1,7 +1,7 @@
-package co.touchlab.faktory.artifactmanager
+package co.touchlab.kmmbridge.artifactmanager
 
-import co.touchlab.faktory.capitalized
-import co.touchlab.faktory.publishingExtension
+import co.touchlab.kmmbridge.capitalized
+import co.touchlab.kmmbridge.publishingExtension
 import org.gradle.api.GradleException
 import org.gradle.api.Project
 import org.gradle.api.Task
