+++ kmmbridge/src/main/kotlin/co/touchlab/kmmbridge/dependencymanager/SpmDependencyManager.kt
@@ -11,18 +11,18 @@
  * the License.
  */
 
-package co.touchlab.faktory.dependencymanager
+package co.touchlab.kmmbridge.dependencymanager
 
-import co.touchlab.faktory.TASK_GROUP_NAME
-import co.touchlab.faktory.domain.SwiftToolVersion
-import co.touchlab.faktory.domain.TargetPlatform
-import co.touchlab.faktory.dsl.TargetPlatformDsl
-import co.touchlab.faktory.findXCFrameworkAssembleTask
-import co.touchlab.faktory.kmmBridgeExtension
-import co.touchlab.faktory.kotlin
-import co.touchlab.faktory.layoutBuildDir
-import co.touchlab.faktory.urlFile
-import co.touchlab.faktory.zipFilePath
+import co.touchlab.kmmbridge.TASK_GROUP_NAME
+import co.touchlab.kmmbridge.domain.SwiftToolVersion
+import co.touchlab.kmmbridge.domain.TargetPlatform
+import co.touchlab.kmmbridge.dsl.TargetPlatformDsl
+import co.touchlab.kmmbridge.findXCFrameworkAssembleTask
+import co.touchlab.kmmbridge.kmmBridgeExtension
+import co.touchlab.kmmbridge.kotlin
+import co.touchlab.kmmbridge.layoutBuildDir
+import co.touchlab.kmmbridge.urlFile
+import co.touchlab.kmmbridge.zipFilePath
 import domain.konanTarget
 import domain.swiftPackagePlatformName
 import org.gradle.api.Action
