+++ website/versioned_docs/version-0.3.x/DEFAULT_GITHUB_FLOW.md
@@ -65,7 +65,7 @@ Find the `build.gradle.kts` file where you configure the multiplatform module yo
 ```kotlin
 plugins {
     kotlin("multiplatform")
-    id("co.touchlab.faktory.kmmbridge") version "{{VERSION_NAME_3X}}"
+    id("co.touchlab.kmmbridge.kmmbridge") version "{{VERSION_NAME_3X}}"
     `maven-publish`
 }
 ```
